+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   * Updated Spotifyd to 357e1476fd4e987b82d00ac7b70891d1730509e8
   * Rewrote MPRIS interface to no longer poll (used by spotify and airplay)
   * Fixed bug in spotify with metadata file location
+  * Fixed Spotify stream field validation to permit spaces
+
 ## 0.3.6
 * Web App
   * Add initial support for song browser (only Pandora stations for now)